Put all mysql config into a temporary option file.
This is (I think) a better approach towards dealing with #129.  I have
tested it a little bit, and I think it may provide a basis for dealing
with #108.  One plus is that it doesn't have to read anything magical
from stderr.

The Tempfile docs say that the file is created 0600.  I think I read
that MySQL ignores option files that are not 0600 or 0400.  I would
rather unlink the option file sooner, but I'm not sure how to do that
without creating a race condition.

I ran 'DB=mysql2 rake test' using ruby-1.9.3p392 and the only failures
I got had to do with with nokogiri.  I tested a little on my own use
case, just to make sure it looked like it was working.